requirements.txt
# File: requirements.txt

# Core Dependencies
faster-whisper==1.0.3
torch==2.1.2
torchaudio==2.1.2

# Video Processing
ffmpeg-python==0.2.0
moviepy==1.0.3
opencv-python==4.9.0.80

# NLP & Embeddings
sentence-transformers==2.3.1
transformers==4.36.2
langchain==0.1.4
langchain-community==0.0.13

# Ollama Integration
ollama==0.1.6

# API Integrations
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
requests-oauthlib==1.3.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
pydantic==2.5.3
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==7.4.4
pytest-cov==4.1.0